+++ test/lib/bats-setup.bash
@@ -18,4 +18,5 @@ set +o 'nounset'
 
 # Do not set/unset: 'keyword', 'noexec'
 
+# Use 6 to enable bashlib64 function debug
 export BL64_LIB_DEBUG=0